Various miscellaneous contributions from AnErrupTion (#993)
* Squashed commit of the following:

commit dd13888720fdbc0651287388a812b9ed4b44a333
Author: AnErrupTion <anerruption@disroot.org>
Date:   Tue Nov 29 17:44:58 2022 +0100

    Remove CommandLine.dll in Mosa.Tools.Package.nuspec

commit 0e9a79eb442dbf7ddbc687ff04913f6a78298b74
Author: AnErrupTion <anerruption@disroot.org>
Date:   Tue Nov 29 17:37:41 2022 +0100

    Remove file copying process for non-existent CommandLine.dll in installer script

commit b2adef80034eded22b2db208b9d82c6cd23101cf
Merge: cdc3b69c6 681a47f57
Author: AnErrupTion <anerruption@disroot.org>
Date:   Tue Nov 29 17:20:34 2022 +0100

    Merge master into branch

commit cdc3b69c64e8e3700c040f436e228386b2272474
Author: AnErrupTion <anerruption@disroot.org>
Date:   Tue Nov 29 16:58:28 2022 +0100

    Add progress bar

commit c7a21273d214b03d447a43af10b721373a7561f6
Author: AnErrupTion <anerruption@disroot.org>
Date:   Fri Oct 14 21:43:47 2022 +0200

    Try to fix Windows build workflow

commit 6795dcb5831a175256a84942167a867ddf342121
Author: AnErrupTion <anerruption@disroot.org>
Date:   Fri Oct 14 21:34:14 2022 +0200

    Make some changes to the GUI launcher

commit 05bdb569f6fa65ebba25282f94d24e016d796264
Author: AnErrupTion <anerruption@disroot.org>
Date:   Fri Oct 14 07:56:48 2022 +0200

    Changes in Internal.MemoryCopy()

commit edb35f90acadede8078a0fdfd37c2bf4922a3870
Author: AnErrupTion <anerruption@disroot.org>
Date:   Wed Oct 12 17:13:33 2022 +0200

    Revert "-enable-kvm" change

commit 748bf086b6fc504baee8e7f6dce809bdc2917bc3
Author: AnErrupTion <anerruption@disroot.org>
Date:   Wed Oct 12 16:52:32 2022 +0200

    Add demo files

commit a352a57136eb712292910910c4ee0651079b75eb
Author: AnErrupTion <anerruption@disroot.org>
Date:   Wed Oct 12 16:51:13 2022 +0200

    Various miscellaneous changes

# Conflicts:
#	Source/Mosa.Tool.Launcher/MainWindow.axaml.cs
#	Source/Mosa.Utility.Launcher/Starter.cs

* - Added transformation types

* - Added transformation types

* - Updates

* - Updates

* - Updates

* - Updates

* - Updates

* - Updates

* - Updates

* - Updates

* - Updates

* - Updates

* - Updates

* - Updates

+++ Source/Mosa.Compiler.Framework/MosaTypeLayout.cs
@@ -1,12 +1,11 @@
 // Copyright (c) MOSA Project. Licensed under the New BSD License.
 
-using Mosa.Compiler.Common;
-using Mosa.Compiler.MosaTypeSystem;
-
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
+using Mosa.Compiler.Common;
+using Mosa.Compiler.MosaTypeSystem;
 
 namespace Mosa.Compiler.Framework
 {
@@ -188,7 +187,7 @@ namespace Mosa.Compiler.Framework
 			lock (_lock)
 			{
 				ResolveType(method.DeclaringType);
-				return (uint)methodSlots[method];
+				return methodSlots[method];
 			}
 		}
